requirements.txt
alembic==1.4.2
colorama==0.4.3
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
pdfkit==0.6.1
pickleshare==0.7.4
Pillow==7.2.0
psycopg2==2.8.6
python-dotenv==0.14.0
python-editor==1.0.4
python-pdf==0.37
requests==2.24.0
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.8
wkhtmltopdf==0.2
WTForms==2.3.1
